+++ Assets/_KYUB/KyubEngine.lib/Kyub.UICommons/Runtime/Scripts/NativeIF/Plugins/NiceJson.cs
@@ -34,7 +34,7 @@ using System.Collections;
 using System.Globalization;
 using System.Text;
 
-namespace MobileInputNativePlugin.NiceJson
+namespace Kyub.Internal.NativeInputPlugin.NiceJson
 {
 	[Serializable]
 	public abstract class JsonNode