+++ NativePlugins/NativeInputField/mobileinput/src/main/java/kyub/uicommons/mobileinput/KeyboardObserver.java
@@ -1,0 +1,9 @@
+package kyub.uicommons.mobileinput;
+
+public interface KeyboardObserver {
+    void onKeyboardHeight(float height, int keyboardHeight, int orientation);
+}
+
+
+
+